Add native Linux support with an SDL2 frontend
The emulation core now builds on Linux (and other non-Windows platforms)
alongside a new SDL2 frontend, while the Windows build keeps using
Win32/DirectDraw/DirectInput/FMOD exactly as before.

Core portability:
- Add src/platform/compat.h, which includes windows.h on Windows and
  otherwise provides the Win32 bits the core uses (wide-filename fopen/stat,
  current directory handling, wsprintfW with Win32 %s semantics, Sleep)
  on top of POSIX, plus UTF-8 conversion helpers for all platforms
- Use a 32-bit DWORD pixel type everywhere (unsigned long is 64 bits on
  LP64 Linux), and fix copy_line32's memcpy size to match
- Only use the x86 inline assembly in the renderer and scale2x MMX filter
  on 32-bit x86, falling back to the existing C code elsewhere
- Fix the case of the renderer.h include and path separator handling
- Move the keyboard-to-joypad mapping from directinput.cpp to a shared
  input.cpp, the FMOD output code to sound_fmod.cpp behind a small
  sound_output_write/reset interface, and the shared globals from main.cpp
  to globals.cpp, so every frontend can reuse them
- Load libgblink with dlopen on non-Windows systems

SDL2 frontend (src/platform/sdl):
- Renderer with the same interface as the DirectDraw one, including frame
  mixing, filters, SGB borders, two-GB mode, visual rumble and on-screen
  messages (with a small built-in font)
- Keyboard input translated to DirectInput key codes, so hhugboy.cfg is
  the same format on both platforms, plus joystick support
- Queued SDL audio, which also paces emulation while sound is on
- Same hotkeys as Windows, drag and drop, and command line options for
  debugging and automated testing (--frames, --screenshot, --unthrottled,
  --debug-log, --no-dialogs)

CMake now splits the sources into core, Windows and SDL groups and finds
SDL2, libpng and zlib on non-Windows systems. The readme has build and
usage instructions for Linux.

+++ src/SGB.cpp
@@ -816,7 +816,7 @@ void sgb_execute_command()
                   copy_line16((unsigned short*)GB->gfx_buffer,(unsigned short*)GB->gfx_buffer_old,160*144);
                } else
                {
-                  copy_line32((unsigned long*)GB->gfx_buffer,(unsigned long*)GB->gfx_buffer_old,160*144);
+                  copy_line32((DWORD*)GB->gfx_buffer,(DWORD*)GB->gfx_buffer_old,160*144);
                }
 
                (renderer.*renderer.drawScreen)();
